Improve Form::error method
Previously the method failed if there were errors but no error for
the given key (because index 0 did not exist). Now an empty array
is returned if there is no error for the given key.

If no key is provided the method will now return the first error
no matter what the keys of the errors are or an empty array if
there are no errors at all.